requirements.txt
Django==2.2.8
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
gunicorn==20.0.2
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
